+++ packages/models/src/index.ts
@@ -1,0 +1,5 @@
+export { IMAGE_MODELS } from "./image-models.generated.js";
+export * from "./image-models.js";
+export { MODELS } from "./models.generated.js";
+export * from "./models.js";
+export * from "./types.js";